IOS-3381 Add builder
